requirements.txt
altair==4.1.0
azureml==0.2.7
azureml_train_restclients_hyperdrive==1.32.0
inference_schema==1.3.0
joblib==1.0.1
numpy==1.19.5
pandas==1.2.4
Pillow==9.0.0
plotly==5.1.0
PyMySQL==1.0.2
scikit_learn==1.0.2
streamlit==1.3.0
